requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
pymongo==4.7.3
bcrypt==4.1.3
dnspython==2.6.1
python-dotenv==1.0.1
flask-cors==4.0.0
razorpay==1.2.0
